Backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sentence-transformers==4.1.0
upstash-vector==0.8.0
SQLAlchemy==2.0.40
psycopg2-binary==2.9.10
python-dotenv==1.1.0
passlib[bcrypt]==1.7.4
python-jose==3.4.0
redis==6.0.0
pydantic==2.11.4
pydantic[email]==2.11.4